+++ R/recode_into.r
@@ -15,6 +15,11 @@
 #' @param default Indicates the default value that is chosen when no match in
 #' the formulas in `...` is found. If not provided, `NA` is used as default
 #' value.
+#' @param overwrite Logical, if `TRUE` (default) and more than one recode pattern
+#' apply to the same case, already recoded values will be overwritten by subsequent
+#' recode patterns. If `FALSE`, former recoded cases will not be altered by later
+#' recode patterns that would apply to those cases again. A warning message is
+#' printed to alert such situations and to avoid unintentional recodings.
 #' @param verbose Toggle warnings.
 #'
 #' @return A vector with recoded values.
@@ -25,6 +30,25 @@
 #'   x > 15 ~ "a",
 #'   x > 10 & x <= 15 ~ "b",
 #'   default = "c"
+#' )
+#'
+#' x <- 1:10
+#' # default behaviour: second recode pattern "x > 5" overwrites
+#' # some of the formerly recoded cases from pattern "x >= 3 & x <= 7"
+#' recode_into(
+#'   x >= 3 & x <= 7 ~ 1,
+#'   x > 5 ~ 2,
+#'   default = 0,
+#'   verbose = FALSE
+#' )
+#'
+#' # setting "overwrite = FALSE" will not alter formerly recoded cases
+#' recode_into(
+#'   x >= 3 & x <= 7 ~ 1,
+#'   x > 5 ~ 2,
+#'   default = 0,
+#'   overwrite = FALSE,
+#'   verbose = FALSE
 #' )
 #'
 #' set.seed(123)
@@ -49,7 +73,7 @@
 #'   default = 0
 #' )
 #' @export
-recode_into <- function(..., data = NULL, default = NA, verbose = TRUE) {
+recode_into <- function(..., data = NULL, default = NA, overwrite = TRUE, verbose = TRUE) {
   dots <- list(...)
 
   # get length of vector, so we know the length of the output vector
@@ -117,15 +141,28 @@ recode_into <- function(..., data = NULL, default = NA, verbose = TRUE) {
     } else {
       already_exists <- out[index] != default
     }
+    # save indices of overwritten cases
+    overwritten_cases <- which(index)[already_exists]
+    # tell user...
     if (any(already_exists) && verbose) {
-      insight::format_warning(
-        paste(
+      if (overwrite) {
+        msg <- paste(
           "Several recode patterns apply to the same cases.",
           "Some of the already recoded cases will be overwritten with new values again",
-          sprintf("(e.g. pattern %i overwrites the former recode of case %i).", i, which(already_exists)[1])
-        ),
-        "Please check if this is intentional!"
-      )
+          sprintf("(e.g. pattern %i overwrites the former recode of case %i).", i, overwritten_cases[1])
+        )
+      } else {
+        msg <- paste(
+          "Several recode patterns apply to the same cases.",
+          "Some of the already recoded cases will not be altered by later recode patterns.",
+          sprintf("(e.g. pattern %i also matches the former recode of case %i).", i, overwritten_cases[1])
+        )
+      }
+      insight::format_warning(msg, "Please check if this is intentional!")
+    }
+    # if user doesn't want to overwrite, remove already recoded indices
+    if (!overwrite) {
+      index[overwritten_cases] <- FALSE
     }
     out[index] <- value
   }