+++ tests/testthat/test-data_to_numeric.R
@@ -8,7 +8,9 @@ test_that("convert character to numeric", {
 })
 
 test_that("convert character to numeric Date", {
-  expect_warning(expect_identical(to_numeric(as.Date("2022-01-01")), 18993))
+  expect_warning(expect_identical(to_numeric(as.Date("2022-01-01")), as.numeric(as.Date("2022-01-01"))))
+  expect_warning(expect_identical(to_numeric(as.POSIXct("2022-01-01")), as.numeric(as.POSIXct("2022-01-01"))))
+  expect_warning(expect_identical(to_numeric(as.POSIXlt("2022-01-01")), as.numeric(as.POSIXlt("2022-01-01"))))
 })
 
 test_that("convert character to numeric preserve levels", {