+++ tests/testthat/test-standardize_models.R
@@ -204,8 +204,6 @@ test_that("variables evaluated in the environment", {
   w <- capture_warnings(standardize(m))
   expect_true(any(grepl("mtcars$mpg", w, fixed = TRUE)))
 
-
-  skip_if(packageVersion("base") == package_version(3.4))
   ## Note:
   # No idea why this is suddenly not giving a warning on older R versions.
   m <- lm(mtcars$mpg ~ mtcars$cyl + mtcars$am, data = mtcars)