+++ NEWS.md
@@ -10,6 +10,9 @@ BUG FIXES
 * Fixed issues in `data_write()` when writing labelled data into SPSS format
   and vectors were of different type as value labels.
 
+* Fixed issue in `recode_into()` with probably wrong case number printed in the
+  warning when several recode patterns match to one case.
+
 # datawizard 0.8.0
 
 BREAKING CHANGES