+++ NEWS.md
@@ -13,6 +13,10 @@ BUG FIXES
 * Fixed issue in `recode_into()` with probably wrong case number printed in the
   warning when several recode patterns match to one case.
 
+* Fixed issue in `data_filter()` where functions containing a `=` (e.g. when
+  naming arguments, like `grepl(pattern, x = a)`) were mistakenly seen as
+  faulty syntax. 
+
 # datawizard 0.8.0
 
 BREAKING CHANGES