+++ NEWS.md
@@ -2,7 +2,7 @@
 
 NEW FUNCTIONS
 
-* `row_means()`, to compute row means, optionally only for those row with at
+* `row_means()`, to compute row means, optionally only for the rows with at
   least `min_valid` non-missing values.
 
 * `means_by_group()`, to compute mean values of variables, grouped by levels